Migrate more packs endpoints
