+++ server/service/endpoint_packs.go
@@ -48,77 +48,6 @@ func packResponseForPack(ctx context.Context, svc fleet.Service, pack fleet.Pack
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-// Delete Pack By ID
-////////////////////////////////////////////////////////////////////////////////
-
-type deletePackByIDRequest struct {
-	ID uint
-}
-
-type deletePackByIDResponse struct {
-	Err error `json:"error,omitempty"`
-}
-
-func (r deletePackByIDResponse) error() error { return r.Err }
-
-func makeDeletePackByIDEndpoint(svc fleet.Service) endpoint.Endpoint {
-	return func(ctx context.Context, request interface{}) (interface{}, error) {
-		req := request.(deletePackByIDRequest)
-		err := svc.DeletePackByID(ctx, req.ID)
-		if err != nil {
-			return deletePackByIDResponse{Err: err}, nil
-		}
-		return deletePackByIDResponse{}, nil
-	}
-}
-
-////////////////////////////////////////////////////////////////////////////////
-// Apply Pack Spec
-////////////////////////////////////////////////////////////////////////////////
-
-type applyPackSpecsRequest struct {
-	Specs []*fleet.PackSpec `json:"specs"`
-}
-
-type applyPackSpecsResponse struct {
-	Err error `json:"error,omitempty"`
-}
-
-func (r applyPackSpecsResponse) error() error { return r.Err }
-
-func makeApplyPackSpecsEndpoint(svc fleet.Service) endpoint.Endpoint {
-	return func(ctx context.Context, request interface{}) (interface{}, error) {
-		req := request.(applyPackSpecsRequest)
-		_, err := svc.ApplyPackSpecs(ctx, req.Specs)
-		if err != nil {
-			return applyPackSpecsResponse{Err: err}, nil
-		}
-		return applyPackSpecsResponse{}, nil
-	}
-}
-
-////////////////////////////////////////////////////////////////////////////////
-// Get Pack Spec
-////////////////////////////////////////////////////////////////////////////////
-
-type getPackSpecsResponse struct {
-	Specs []*fleet.PackSpec `json:"specs"`
-	Err   error             `json:"error,omitempty"`
-}
-
-func (r getPackSpecsResponse) error() error { return r.Err }
-
-func makeGetPackSpecsEndpoint(svc fleet.Service) endpoint.Endpoint {
-	return func(ctx context.Context, request interface{}) (interface{}, error) {
-		specs, err := svc.GetPackSpecs(ctx)
-		if err != nil {
-			return getPackSpecsResponse{Err: err}, nil
-		}
-		return getPackSpecsResponse{Specs: specs}, nil
-	}
-}
-
-////////////////////////////////////////////////////////////////////////////////
 // Get Pack Spec
 ////////////////////////////////////////////////////////////////////////////////
