+++ server/service/service_packs_test.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"testing"
 
-	"github.com/fleetdm/fleet/v4/server/datastore/mysql"
 	"github.com/fleetdm/fleet/v4/server/fleet"
 	"github.com/fleetdm/fleet/v4/server/mock"
 	"github.com/fleetdm/fleet/v4/server/ptr"
@@ -38,132 +37,4 @@ func TestNewSavesTargets(t *testing.T) {
 	assert.Equal(t, uint(123), pack.HostIDs[0])
 	assert.Equal(t, uint(456), pack.LabelIDs[0])
 	assert.Equal(t, uint(789), pack.TeamIDs[0])
-}
-
-func TestService_DeletePackByID_GlobalPack(t *testing.T) {
-	ds := mysql.CreateMySQLDS(t)
-	defer ds.Close()
-	test.AddAllHostsLabel(t, ds)
-
-	globalPack, err := ds.EnsureGlobalPack(context.Background())
-	require.NoError(t, err)
-
-	type fields struct {
-		ds fleet.Datastore
-	}
-	type args struct {
-		ctx context.Context
-		id  uint
-	}
-	tests := []struct {
-		name    string
-		fields  fields
-		args    args
-		wantErr bool
-	}{
-		{
-			name: "cannot delete global pack",
-			fields: fields{
-				ds,
-			},
-			args: args{
-				ctx: test.UserContext(test.UserAdmin),
-				id:  globalPack.ID,
-			},
-			wantErr: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			svc := newTestService(tt.fields.ds, nil, nil)
-			if err := svc.DeletePackByID(tt.args.ctx, tt.args.id); (err != nil) != tt.wantErr {
-				t.Errorf("DeletePackByID() error = %v, wantErr %v", err, tt.wantErr)
-			}
-		})
-	}
-}
-
-func TestService_ApplyPackSpecs(t *testing.T) {
-	ds := mysql.CreateMySQLDS(t)
-	defer ds.Close()
-	test.AddAllHostsLabel(t, ds)
-
-	global, err := ds.EnsureGlobalPack(context.Background())
-	require.NoError(t, err)
-
-	users := createTestUsers(t, ds)
-	user := users["admin1@example.com"]
-
-	team1, err := ds.NewTeam(context.Background(), &fleet.Team{
-		ID:          42,
-		Name:        "team1",
-		Description: "desc team1",
-	})
-	require.NoError(t, err)
-
-	teamPack, err := ds.EnsureTeamPack(context.Background(), team1.ID)
-	require.NoError(t, err)
-
-	type fields struct {
-		ds fleet.Datastore
-	}
-	type args struct {
-		ctx   context.Context
-		specs []*fleet.PackSpec
-	}
-	tests := []struct {
-		name    string
-		fields  fields
-		args    args
-		want    []*fleet.PackSpec
-		wantErr bool
-	}{
-		{
-			name: "cannot modify global pack",
-			fields: fields{
-				ds,
-			},
-			args: args{
-				ctx: test.UserContext(&user),
-				specs: []*fleet.PackSpec{
-					{Name: global.Name, Description: "bar", Platform: "baz"},
-					{Name: "Foo Pack", Description: "Foo Desc", Platform: "MacOS"},
-					{Name: "Bar Pack", Description: "Bar Desc", Platform: "MacOS"},
-				},
-			},
-			want: []*fleet.PackSpec{
-				{Name: "Foo Pack", Description: "Foo Desc", Platform: "MacOS"},
-				{Name: "Bar Pack", Description: "Bar Desc", Platform: "MacOS"},
-			},
-			wantErr: false,
-		},
-		{
-			name: "cannot modify team pack",
-			fields: fields{
-				ds,
-			},
-			args: args{
-				ctx: test.UserContext(&user),
-				specs: []*fleet.PackSpec{
-					{Name: teamPack.Name, Description: "Desc", Platform: "windows"},
-					{Name: "Test", Description: "Test Desc", Platform: "linux"},
-				},
-			},
-			want: []*fleet.PackSpec{
-				{Name: "Test", Description: "Test Desc", Platform: "linux"},
-			},
-			wantErr: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			svc := newTestService(tt.fields.ds, nil, nil)
-			got, err := svc.ApplyPackSpecs(tt.args.ctx, tt.args.specs)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("ApplyPackSpecs() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			require.Equal(t, tt.want, got)
-		})
-	}
 }
